+++ CHANGELOG.md
@@ -9,6 +9,15 @@ All notable changes to Rainbow-Node-SDK will be documented in this file.
 ## [1.70.0-dotnet.0] - 2020-04-XX
 -   Add defaultDEBUG to generate the compiled files with debug log not putted in comment.
 -   Fix generateRss file because Array.values() method did not work anymore.
+-   Update getContactByJid with a forceServerSearch parameter to force the search of the _contacts informations on the server.
+-   Add method `channelsService::likeItem` to like treatment/event on a Channel's Item with an appreciation
+-   Add method `channelsService::getDetailedAppreciations` to know in details apprecations given on a channel item (by userId the apprecation given)
+-   Add like events on a Channel's Item with an appreciation.
+-   Add in `postChangeLogInChannel` code to like the item when posting the changelog in Rainbow CPaaS info channel.
+-   Fix `postChangeLogInChannel` to wait until the publish and like item are done.
+-   Fix treatment of XMPP iq query event `set "remove"` from roster which was sending an "unavailable service" to server.
+-   Add method `contactsService::removeFromNetwork` to remove a contact from the list of contacts and unsubscribe to the contact's presence
+-   Add event `rainbow_contactremovedfromnetwork` raised when a contact is removed from connected user's network.
 
 ## [1.69.0] - 2020-03-30
 -   Add nextDebugBuildVersion program to generate to stdout a new debug version of  rainbow-node-sdk